use native timestamps for watchers

+++ ozobot-ari/src/ozobot/ari/webprotocol/types.py
@@ -15,4 +15,4 @@ class PlayAudioRequest(rpctypes.BaseRequest):
 class TimeOfFlightResponse(rpctypes.BaseResponse):
     distance: float
     deviation: float
-    timestamp: float
+    timestamp: int
